Move the options and sanitizers for the rules settings into the RulesPage class directly
